Setup Changelog

+++ Makefile
@@ -19,8 +19,16 @@ build: deps
 	tar -zcf release/$(NAME)_darwin_$(ARCH).tgz -C build/Darwin $(NAME)
 
 release:
+ifeq ($(shell git diff --shortstat 2> /dev/null | tail -n1),)
+	git push
+	github-changes -o webminal -r webminal --use-commit-body --no-merges
+	git add CHANGELOG.md
+	git commit -m "Update CHANGELOG.md"
 	git tag -f -s v$(VERSION)
 	git push --tags --force
+else
+	@echo "Please cleanup working directory." && exit 1
+endif
 
 test: deps
 	go test ./...
